+++ engine/src/main/java/nl/inl/blacklab/resultproperty/HitPropertyHitText.java
@@ -65,7 +65,7 @@ public class HitPropertyHitText extends HitPropertyContextBase {
     }
 
     @Override
-    public PropertyValueContextWords get(int hitIndex) {
+    public PropertyValueContextWords get(long hitIndex) {
         int[] context = contexts.get(hitIndex);
         int contextHitStart = context[Contexts.HIT_START_INDEX];
         int contextRightStart = context[Contexts.RIGHT_START_INDEX];
@@ -82,7 +82,7 @@ public class HitPropertyHitText extends HitPropertyContextBase {
     }
 
     @Override
-    public int compare(int indexA, int indexB) {
+    public int compare(long indexA, long indexB) {
         int[] ca = contexts.get(indexA);
         int caHitStart = ca[Contexts.HIT_START_INDEX];
         int caRightStart = ca[Contexts.RIGHT_START_INDEX];
