+++ engine/src/main/java/nl/inl/blacklab/resultproperty/HitPropertyWordLeft.java
@@ -62,7 +62,7 @@ public class HitPropertyWordLeft extends HitPropertyContextBase {
     }
 
     @Override
-    public PropertyValueContextWord get(int hitIndex) {
+    public PropertyValueContextWord get(long hitIndex) {
         int[] context = contexts.get(hitIndex);
         int contextHitStart = context[Contexts.HIT_START_INDEX];
         //int contextRightStart = context[Contexts.CONTEXTS_RIGHT_START_INDEX];
@@ -76,7 +76,7 @@ public class HitPropertyWordLeft extends HitPropertyContextBase {
     }
 
     @Override
-    public int compare(int indexA, int indexB) {
+    public int compare(long indexA, long indexB) {
         int[] ca = contexts.get(indexA);
         int caHitStart = ca[Contexts.HIT_START_INDEX];
         int caLength = ca[Contexts.LENGTH_INDEX];
