+++ server/src/main/java/nl/inl/blacklab/server/requesthandlers/RequestHandler.java
@@ -834,14 +834,14 @@ public abstract class RequestHandler {
             // Viewing single group of documents, possibly based on a hits search.
             // group.getResults().getOriginalHits() returns null in this case,
             // so we have to iterate over the DocResults and sum up the hits ourselves.
-            int numberOfHits = 0;
+            long numberOfHits = 0;
             for (DocResult dr : docResults) {
                 numberOfHits += dr.size();
             }
             ds.entry("numberOfHits", numberOfHits)
                     .entry("numberOfHitsRetrieved", numberOfHits);
 
-            int numberOfDocsCounted = docResults.size();
+            long numberOfDocsCounted = docResults.size();
             if (countFailed)
                 numberOfDocsCounted = -1;
             ds.entry("numberOfDocs", numberOfDocsCounted)
@@ -946,7 +946,7 @@ public abstract class RequestHandler {
         addSummaryCsvCommon(printer, numColumns, searchParam, groups, subcorpusSize);
 
         writeRow(printer, numColumns, "summary.numberOfDocs", docResults.size());
-        writeRow(printer, numColumns, "summary.numberOfHits", docResults.stream().collect(Collectors.summingInt(r -> r.size())));
+        writeRow(printer, numColumns, "summary.numberOfHits", docResults.stream().collect(Collectors.summingLong(r -> r.size())));
     }
 
     protected static BlsException translateSearchException(Exception e) {
