+++ server/src/main/java/nl/inl/blacklab/server/requesthandlers/RequestHandlerDocs.java
@@ -113,10 +113,10 @@ public class RequestHandlerDocs extends RequestHandler {
         if (sortProp != null)
             docsSorted = docsSorted.sort(sortProp);
 
-        int first = searchParam.getInteger("first");
+        long first = searchParam.getLong("first");
         if (first < 0)
             first = 0;
-        int number = searchParam.getInteger("number");
+        long number = searchParam.getLong("number");
         if (number < 0 || number > searchMan.config().getParameters().getPageSize().getMax())
             number = searchMan.config().getParameters().getPageSize().getDefaultValue();
         totalDocResults = docsSorted;
@@ -202,7 +202,7 @@ public class RequestHandlerDocs extends RequestHandler {
 
             // Combine all
             ds.entry("docPid", pid);
-            int numHits = result.size();
+            long numHits = result.size();
             if (numHits > 0)
                 ds.entry("numberOfHits", numHits);
